refactor(services): extract web session recovery and chat config state
Takes the two bindings the chat cut freed. recoveryReportsBySession
moves into web-session-recovery-state.ts with the report factory, the
summary read and the two exported recovery seams; memoryChatConfigs
moves into web-chat-config-state.ts with its storage read/write pair.
Neither binding is exported.

Five methods move: the recovery trio into web-session-recovery-client.ts
over SessionRecoveryService, and the two session chat-config methods
into web-session-chat-config-client.ts over SessionChatConfigService.
Both interfaces' signatures moved out of AgentService rather than being
copied.

deleteSession stays in the composition root for now and reaches both
through deleteWebRecoveryReports and deleteWebSessionChatConfig, the
latter replacing an inline copy-delete-write of the config map.

web-agent-client.ts 1514 -> 1332, agent-service.ts 342 -> 338.
Subtree aggregate 19203 -> 19286.

tsc --noEmit, contracts:check and the full suite pass: 286 files /
1301 tests, unchanged, with no test file edited.
